+++ lib/esse/rspec.rb
@@ -3,8 +3,13 @@
 require "esse"
 
 require_relative "rspec/version"
+require_relative "rspec/class_methods"
 
 module Esse
   module RSpec
   end
+end
+
+::RSpec.configure do |config|
+  config.include Esse::RSpec::ClassMethods
 end