+++ CHANGELOG.md
@@ -4,5 +4,8 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.0.1 - 2023-11-10
+## 0.0.1 - 2023-11-15
 The first release of the Esse::RSpec
+* stub_esse_class helper
+* stub_esse_index helper
+* stub_esse_search helper